Made it so you can now get active posts
You can now get active posts for weather, traffic, and both of those
combined. You get the posts with an http get request to:

localhost:3000/api/traffic
localhost:3000/api/weather
localhost:3000/api/allPosts

There are now copy cat models for 'active' models like ActivePost,
ActiveWeather, and ActiveTraffic. These are exact copies of the old
models except for the names. These are emptied in every new cron timer
cycle and replenished with whatever is current. This allows us to have
access to the current posts on the front end.

The old models and posts will be kept and logged. They are in a folder
called loggedAlerts. These collections are still checked to see if an
alert already exists before they are added.