I fixed the gulp problem and got the data to display.
The gulp issue was simply a missing '})' in app_client/main.js. Gulp now
works perfectly and minifies the controllers and main.js. To run the
command enter: gulp in the command line in the main project directory.

I made a few changes to your controller rob that made the getPosts
function work. All I had to do was change what the postData function
returned. This makes it so the data now displays in the front end. We
still have to parse and seperate it but we're moving forward now.

Also I changed the function call WeatherCtrl to weatherCtrl in the
/app_client/weather folder because that was causing issues.