Created substring on backend
I consolidated the models into one place

I updated the user schema with filters and deviceId

I got the brief description done on the backend and posted it on the
front end. I think we might want to do that substring on the front end
though now that I think about it. I'll talk to you guys about that.