I added the clickMore variable to the post models
The clickMore variable only exists if the post description is greater
than 200 characters. This allows the front end to add a 'click more'
style button when the description is greater than 200 characters, and
leave it blank when it is smaller.

I updater the weather and traffic functions to support this.