Filters!
Filtering functionality is done, we just need a page or something where
we can set filters easily.

Right now if you want to change the filter to test it yourself you will
have to change the filter array when a user is registered. If you want
both to show up, put in both "weather" and "traffic", if you want
nothing, keep the array empty.